// src/app/interfaces/icliente-edit.ts
export interface IClienteEdit {
  cidadeId: number;
  representanteId: number;
  documento: string; // ok
  razaoSocial: string; // ok
  nomeFantasia: string; // ok
  endereco: string; // ok
  bairro: string; // ok
  complemento: string; // ok
  cep: string; // ok
  ddd: string; // ok
  telefone: string; //ok
  email: string; // ok
  inscricaoEstadual: string; //ok
  pessoa: string; // ok
  avaliacao: string; //ok
  limiteDeCredito: number;
  inscricaoMunicipal: string;
  contato: string;
  suframa: string;
}